[MDC-654] nmap: nping: set cap_net_raw+ep or make it setuid
If setcap is available (i.e. libcap2-bin is unpacked), use it to give
nping the cap_net_raw capabilities, which allows it to work well when
called with --privileged.  Otherwise, make it setuid root.

Since the package will now force nping to be either CAP_NET_RAW=ep or
setuid(0), it should assume it needs to run in --privileged mode.
Patch it accordingly.